+++ designer/client/src/containers/muiThemeProvider.tsx
@@ -1,9 +1,0 @@
-import { ThemeProvider } from "@mui/material/styles";
-import React, { PropsWithChildren } from "react";
-import { useMuiTheme } from "./useMuiTheme";
-
-export const MuiThemeProvider: React.FC<PropsWithChildren<unknown>> = ({ children }) => {
-    const muiTheme = useMuiTheme();
-
-    return <ThemeProvider theme={muiTheme}>{children}</ThemeProvider>;
-};