+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/ComponentIdProvider.scala
@@ -1,47 +1,22 @@
-package pl.touk.nussknacker.ui.component
+package pl.touk.nussknacker.engine.definition
 
-import com.typesafe.scalalogging.LazyLogging
-import pl.touk.nussknacker.engine.ProcessingTypeData
 import pl.touk.nussknacker.engine.api.component.ComponentType.ComponentType
 import pl.touk.nussknacker.engine.api.component.{ComponentId, ComponentType}
 import pl.touk.nussknacker.engine.component.ComponentUtil
 import pl.touk.nussknacker.engine.component.ComponentsUiConfigExtractor.ComponentsUiConfig
 import pl.touk.nussknacker.engine.graph.node.{NodeData, WithComponent}
-import pl.touk.nussknacker.restmodel.process.ProcessingType
-import pl.touk.nussknacker.ui.process.ProcessCategoryService
 
 //TODO: It is work around for components duplication across multiple scenario types, until we figure how to do deduplication.
 trait ComponentIdProvider {
-  def createComponentId(processingType: ProcessingType, name: Option[String], componentType: ComponentType): ComponentId
-  def nodeToComponentId(processingType: ProcessingType, node: NodeData): Option[ComponentId]
+  def createComponentId(processingType: String, name: Option[String], componentType: ComponentType): ComponentId
+  def nodeToComponentId(processingType: String, node: NodeData): Option[ComponentId]
 }
 
-object DefaultComponentIdProvider extends LazyLogging {
-
-  def createUnsafe(
-      processingTypeDataMap: Map[ProcessingType, ProcessingTypeData],
-      categoryService: ProcessCategoryService
-  ): ComponentIdProvider = {
-    logger.debug("Creating component id provider")
-
-    val componentObjectsService = new ComponentObjectsService(categoryService)
-    val componentObjectsMap     = processingTypeDataMap.transform(componentObjectsService.prepareWithoutFragments)
-    val componentIdProvider = new DefaultComponentIdProvider(componentObjectsMap.transform {
-      case (_, componentsObjects) => componentsObjects.config
-    })
-
-    ComponentsValidator.checkUnsafe(componentObjectsMap, componentIdProvider)
-
-    componentIdProvider
-  }
-
-}
-
-class DefaultComponentIdProvider(configs: Map[ProcessingType, ComponentsUiConfig]) extends ComponentIdProvider {
+class DefaultComponentIdProvider(configs: Map[String, ComponentsUiConfig]) extends ComponentIdProvider {
 
   override def createComponentId(
-      processingType: ProcessingType,
-      name: Option[ProcessingType],
+      processingType: String,
+      name: Option[String],
       componentType: ComponentType
   ): ComponentId = {
     name match {
@@ -50,7 +25,7 @@ class DefaultComponentIdProvider(configs: Map[ProcessingType, ComponentsUiConfig
     }
   }
 
-  override def nodeToComponentId(processingType: ProcessingType, node: NodeData): Option[ComponentId] =
+  override def nodeToComponentId(processingType: String, node: NodeData): Option[ComponentId] =
     ComponentUtil
       .extractComponentType(node)
       .map(componentType =>
@@ -60,11 +35,7 @@ class DefaultComponentIdProvider(configs: Map[ProcessingType, ComponentsUiConfig
         }
       )
 
-  private def createComponentId(
-      processingType: ProcessingType,
-      name: String,
-      componentType: ComponentType
-  ): ComponentId = {
+  private def createComponentId(processingType: String, name: String, componentType: ComponentType): ComponentId = {
     val defaultComponentId    = ComponentId.default(processingType, name, componentType)
     val overriddenComponentId = getOverriddenComponentId(processingType, name, defaultComponentId)
 
@@ -81,7 +52,7 @@ class DefaultComponentIdProvider(configs: Map[ProcessingType, ComponentsUiConfig
   }
 
   private def getOverriddenComponentId(
-      processingType: ProcessingType,
+      processingType: String,
       componentName: String,
       defaultComponentId: ComponentId
   ): ComponentId = {